Add files via upload
## Project Structure

instalogin.html: This file contains the HTML structure of the Instagram login page clone. It includes the layout, form elements, and basic content that users interact with when attempting to log in.

style.css-This file contains the CSS styles that define the visual appearance of the login page. It includes styles for layout, colors, fonts, and responsive design to ensure the page looks good on various devices.